There is still one problem

+++ OJ Assignment/Assignment 3/src/ProblemFour.java
@@ -6,40 +6,40 @@ public class ProblemFour {
         Scanner fetch = new Scanner(System.in);
         String initialIntegerInString = fetch.next();
         int segmentLength = fetch.nextInt();
-        int[] theCorrespondingArray = StringIntToArray(initialIntegerInString, segmentLength);
-        int[] theCorrespondingArrayAfterReverse = ReverseAllTheIntegersInTheArray(theCorrespondingArray);
-        int sumOfThePartsOfTheArray = SumUpTheArray(theCorrespondingArrayAfterReverse);
+        long[] theCorrespondingArray = StringIntToArray(initialIntegerInString, segmentLength);
+        long[] theCorrespondingArrayAfterReverse = ReverseAllTheIntegersInTheArray(theCorrespondingArray);
+        long sumOfThePartsOfTheArray = SumUpTheArray(theCorrespondingArrayAfterReverse);
         System.out.println(sumOfThePartsOfTheArray);
     }
 
-    public static int[] StringIntToArray(String theInitialInteger, int segmentLength) {
+    public static long[] StringIntToArray(String theInitialInteger, int segmentLength) {
         int lengthOfInitialNumber = theInitialInteger.length();
         int lengthOfTheArray = (int) Math.ceil((double) lengthOfInitialNumber / (double) segmentLength);
         int errorInLength = lengthOfTheArray * segmentLength - lengthOfInitialNumber;
-        int[] outputArray = new int[lengthOfTheArray];
+        long[] outputArray = new long[lengthOfTheArray];
         for (int computationTime = 0; computationTime < lengthOfTheArray; computationTime++) {
             if (computationTime != lengthOfTheArray - 1) {
-                outputArray[computationTime] = Integer.parseInt(theInitialInteger.substring(computationTime * segmentLength, computationTime * segmentLength + segmentLength));
+                outputArray[computationTime] = Long.parseLong(theInitialInteger.substring(computationTime * segmentLength, computationTime * segmentLength + segmentLength));
             } else {
-                outputArray[computationTime] = Integer.parseInt(theInitialInteger.substring(computationTime * segmentLength, lengthOfInitialNumber));
+                outputArray[computationTime] = Long.parseLong(theInitialInteger.substring(computationTime * segmentLength, lengthOfInitialNumber));
             }
         }
         return outputArray;
     }
 
-    public static int SumUpTheArray(int[] inputArray) {
-        int sumOfTheArray = 0;
+    public static long SumUpTheArray(long[] inputArray) {
+        long sumOfTheArray = 0;
         for (int sequenceOfTheOperationLocation = 0; sequenceOfTheOperationLocation < inputArray.length; sequenceOfTheOperationLocation++) {
             sumOfTheArray += inputArray[sequenceOfTheOperationLocation];
         }
         return sumOfTheArray;
     }
 
-    public static int ReverseAnInteger(int initialInteger) {
+    public static long ReverseAnInteger(long initialInteger) {
         int lengthOfTheInteger = 0;
-        int remainedInteger = initialInteger;
-        int numberToTestLength = initialInteger;
-        int outputInteger = 0;
+        long remainedInteger = initialInteger;
+        long numberToTestLength = initialInteger;
+        long outputInteger = 0;
         while (true) {
             numberToTestLength = numberToTestLength / 10;
             lengthOfTheInteger += 1;
@@ -48,19 +48,19 @@ public class ProblemFour {
             }
         }
         for (int timeOfComputation = 0; timeOfComputation < lengthOfTheInteger; timeOfComputation++) {
-            int computationPart = remainedInteger % 10;
+            long computationPart = remainedInteger % 10;
             outputInteger = outputInteger * 10 + computationPart;
             remainedInteger = remainedInteger / 10;
         }
         return outputInteger;
     }
 
-    public static int[] ReverseAllTheIntegersInTheArray(int[] initialArray) {
+    public static long[] ReverseAllTheIntegersInTheArray(long[] initialArray) {
         int lengthOfTheArray = initialArray.length;
         for (int locationOnTheArray = 0; locationOnTheArray < lengthOfTheArray; locationOnTheArray++) {
             initialArray[locationOnTheArray] = ReverseAnInteger(initialArray[locationOnTheArray]);
         }
-        int[] outputArray = initialArray;
+        long[] outputArray = initialArray;
         return outputArray;
     }
 }
